refactor: Rename test_skip_detector.py to skip_detector.py
The file is a production module that detects test skip patterns, not a
test file itself. The test_ prefix was causing false positives in the
file-placement linter which denies test files in src/.
